requirements.txt
requests_html
BeautifulSoup
pandas
numpy
regex